refactor(tactics): extract useTacticsFilters hook from TacticsTab
Move the player search / position filter state, the filtered starting-XI
and bench memos, and handleClearFilters into a useTacticsFilters hook.
The sortKey/sortDir consts and both memo dependency arrays are preserved
verbatim. Pure move.

TacticsTab.tsx: 750 -> 704 lines.
